+++ packages/design-system/src/components/Page/PageHeader.tsx
@@ -17,7 +17,7 @@ export const PageHeader: FC<PageHeaderProps> = ({ ref, sticky, children, classNa
       data-testid={testId}
       data-slot='page-header'
       className={cn(
-        'flex items-center px-16 pb-16 gap-12',
+        'flex items-center px-16 pb-8 gap-12',
         sticky && 'sticky top-0 z-10 bg-bg-surface-2',
         className,
       )}