+++ packages/design-system/src/components/Pagination/lib/index.ts
@@ -1,0 +1,2 @@
+export { createPageSizeCollection } from './createPageSizeCollection';
+export { useClientPagination } from './useClientPagination';